+++ node.h
@@ -2,10 +2,10 @@
 
 typedef struct underlink_node
 {
-	unsigned long long		nodeID;
+	uint64_t				nodeID;
 	struct sockaddr_in		endpoint;
 	char					publickey[64];
 	int						lastused;
 } underlink_node;
 
-unsigned long long getDistance(underlink_node one, underlink_node two);
+uint64_t getDistance(underlink_node one, underlink_node two);
